upstream(GTK3V2MAIN): sync SF r7293–r7309 (17 changesets)
Auto-synced by xtrkcad-sf-sync on 2026-09-15.
First: fix(gtk3): CleanSegs leaves dangling pointers after freeing memory (SF #746)
Last:  Merge bug-mrulist-orphan-duplicate-gtk3 into GTK3V2MAIN

Hg: r7293..r7309 on GTK3V2MAIN (SF xtrkcad-fork/xtrkcad)

+++ app/wlib/gtk3lib/wrapbox/testwrapbox.c
@@ -74,7 +74,7 @@ static void
 populate_wrapbox_wrappy (EggWrapBox *wrapbox)
 {
 	GtkWidget *widget, *frame;
-	gint i;
+	gulong i;
 
 	const gchar *strings[] = {
 		"These are", "some wrappy label", "texts", "of various", "lengths.",
